Support SCSS
This commit adds support for SCSS and enables hot reloading of CSS
in development mode.

The hot reloading is taken care of by webpack and the `style-loader`
plugin. When the `yarn build` command is used, the `style-loader` isn't
used but instead the `extract-text-plugin` is configured to take all
compiled css chunks into a single .css file.

What this doesn't do:

* deal with de-duplicating any css
* optimising it
* eliminating whitespace,
* any advanced uglyfication
* handle url rewriting