Add way to depend on the result of another step
I have added the concept of `checkApplies` as a possible key in the
step object.

If it is defined then the previous step, when trying to decide where to
send you next, will run that check and pass in the contents of the
cookie.

This check can return false and it will be skipped or true and it will
then be redirected to.