Remove the need for fields object on steps
We currently create an array of objects which has information for
each field used to render them in the view.

Now we are using methods like `.label()` on the Joi objects we can
build that data from those objects and send straight to the views
without having to define those manually on each step.