Get the email address to use for the registration
This gets the practice email address from the environment variables
and uses that as the email send method.

It checks that the practice is in the accepted list and then converts
the key to a constant format so that it can get the value of the
email for that practice.

The email addresses will be stored as environment variables as the
emails provided my not be for public use.